Rewrite FindKernelRandom & immediate verification
Add a helper function
`GenerateRandomKernelElementsAndOptionallyVerifyThem`, which is
used both when computing random generators of a kernel and when doing
immediate verification of a kernel.

Add the function `ImmediateVerification`.

Slightly change the return values of FindKernelRandom and
FindKernelFastNormalClosure: previously `true` denoted success and
`false` denoted that something went wrong while computing an SLP in the
image node. Now in the latter situation `fail` is returned.
This is to bring them in line with the return values of the new function
`ImmediateVerification`, which returns `true` upon successfull
verification, `false` upon finding new kernel generators, and `fail`
upon failure to compute an SLP in the image node.

Also remove the `InfoRecog` progress indicator during immediate
verification.